exlude csv test from bind uploader
